# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Categories Controller",
        "version": "1"
    },
    "paths": {
...
        "/admin/groups/{id}": {
            "get": {
                "tags": [
                    "groups"
                ],
                "operationId": "021c328f28588b06efb6ee2ff53fc1ab",
                "responses": {
                    "204": {
                        "description": "Fetched Group Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminGroup"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "groups"
                ],
                "summary": "Remove the specified resource from storage.",
                "operationId": "dd0b4de268ddb96b71d83abbb8f698bd",
                "responses": {
                    "204": {
                        "description": "The Group was successfully deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "groups"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "c4cddb49dad395ce22690d2fcc9ad32b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminUpdateGroup"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminGroup"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminGroupIdPathParameter"
                }
            ]
        },
...
        "/admin/groups": {
            "get": {
                "tags": [
                    "groups"
                ],
                "summary": "Get All Groups",
                "operationId": "88b2113eba763d404ae87d088c031a8d",
                "responses": {
                    "200": {
                        "description": "The Group was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adminGroup"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "groups"
                ],
                "summary": "Create a new Group.",
                "operationId": "40ad095510ebd19a420248bddef24c63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminCreateGroup"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminGroup"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "adminCreateGroup": {
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "adminGroup": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "default": "2017-02-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    }
                },
                "type": "object"
            },
            "adminUpdateGroup": {
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "adminGroupIdPathParameter": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ]
}